Add an abstract class BaseAuthenticationException
